Update dependencies and add TripAdvisor widget component
- Updated `graphql` to version 16.10.0 in dependencies.
- Updated `@graphql-codegen/cli` to version 5.0.5 in devDependencies.
- Updated `@graphql-codegen/typescript-react-apollo` to version 4.3.2 in devDependencies.
- Added new SCSS file for TripAdvisor widget styling.
- Implemented TripAdvisor widget component with script loading and link to TripAdvisor review.
