+++ app/blog/[slug]/post.module.scss
@@ -1,29 +1,3 @@
-.wrapper {
-  padding: 30px 0;
-}
-
-.row {
-  max-width: 1600px;
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  margin: 0 auto;
-}
-
-.sidebar {
-  flex: 0 0 25%;
-  max-width: 25%;
-  padding: 1rem;
-  background-color: #f9f9f9;
-}
-
-.column {
-  flex: 0 0 75%;
-  max-width: 75%;
-  padding: 0 30px 30px;
-  border-left: 1px solid #ececec;
-}
-
 .article {
   position: relative;
   box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.22);
