--- app/blog/blogComponents/TripAdvisor.module.scss
@@ -1,5 +1,0 @@
-.aside {
-  position: sticky;
-  top: 1rem;
-  border: 1px solid #ececec;
-}
