--- app/blog/blogComponents/TripAdvisor.scss
@@ -1,4 +1,0 @@
-#CDSWIDSSP {
-  width: 100% !important;
-  background-color: yellow;
-}
