+++ app/blog/blogComponents/TripAdvisorWidget.module.scss
@@ -1,0 +1,9 @@
+.tripadvisor-sidebar {
+  position: sticky;
+  top: 1rem;
+
+  .tripadvisor {
+    width: 100% !important;
+    border: 1px solid #ececec;
+  }
+}
