+++ app/blog/layout.tsx
@@ -1,25 +1,20 @@
 import type { Metadata } from 'next';
 import styles from './blog.module.scss';
+import TripAdvisor from './blogComponents/TripAdvisorWidget';
 
 export const metadata: Metadata = {
   title: 'Blog - My Website',
   description: 'Welcome to our blog where we share updates, ideas, and news.',
 };
 
-export default function BlogLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function BlogLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
       {/* Blog Header */}
       <header className={styles.blogHeader}>
         <div className={styles.headerContent}>
           <h1>Welcome</h1>
-          <p className={styles.subHeader}>
-            Activities | Recommendations | Ideas | News
-          </p>
+          <p className={styles.subHeader}>Activities | Recommendations | Ideas | News</p>
           <h2>Let&apos;s dive in!</h2>
         </div>
       </header>
@@ -35,12 +30,16 @@ export default function BlogLayout({
       </div>
 
       {/* Main Blog Content */}
-      <div className={styles.blogContent}>
-        <aside className={styles.tripAdvisorSidebar}>
-          {/* Reserved space for TripAdvisor bar */}
-        </aside>
-        <div className={styles.postsContainer}>{children}</div>
-      </div>
+      <main>
+        <div className='wrapper'>
+          <div className='row'>
+            <aside className='blog sidebar'>
+              <TripAdvisor />
+            </aside>
+            <div className='column'>{children}</div>
+          </div>
+        </div>
+      </main>
     </>
   );
 }
