+++ app/globals.scss
@@ -4,8 +4,36 @@
   box-sizing: border-box;
 }
 
+$border-color: #ececec;
+
 a {
   text-decoration: none;
+}
+
+.wrapper {
+  padding: 30px 0;
+}
+
+.row {
+  max-width: 1600px;
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  margin: 0 auto;
+}
+
+.sidebar {
+  flex: 0 0 25%;
+  max-width: 25%;
+  padding: 1rem;
+  background-color: #f9f9f9;
+}
+
+.column {
+  flex: 0 0 75%;
+  max-width: 75%;
+  padding: 0 30px 30px;
+  border-left: 1px solid $border-color;
 }
 
 /* :root {
