+++ package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@apollo/client": "^3.10.5",
-    "graphql": "^16.8.2",
+    "graphql": "^16.10.0",
     "next": "14.2.3",
     "nextjs-toploader": "^3.7.15",
     "react": "^18",
@@ -23,9 +23,9 @@
     "zustand": "^4.5.4"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^5.0.2",
+    "@graphql-codegen/cli": "^5.0.5",
     "@graphql-codegen/client-preset": "^4.2.6",
-    "@graphql-codegen/typescript-react-apollo": "^4.3.0",
+    "@graphql-codegen/typescript-react-apollo": "^4.3.2",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
